+++ crates/rust-client/src/store/web_store/js/transactions.js
@@ -60,25 +60,15 @@ export async function getTransactions(filter) {
           }
         }
 
-        let inputNotesArrayBuffer =
-          await transactionRecord.inputNotes.arrayBuffer();
-        let inputNotesArray = new Uint8Array(inputNotesArrayBuffer);
-        let inputNotesBase64 = uint8ArrayToBase64(inputNotesArray);
-        transactionRecord.inputNotes = inputNotesBase64;
-
-        let outputNotesArrayBuffer =
-          await transactionRecord.outputNotes.arrayBuffer();
-        let outputNotesArray = new Uint8Array(outputNotesArrayBuffer);
-        let outputNotesBase64 = uint8ArrayToBase64(outputNotesArray);
-        transactionRecord.outputNotes = outputNotesBase64;
+        let metadataArrayBuffer =
+          await transactionRecord.metadata.arrayBuffer();
+        let metadataArray = new Uint8Array(metadataArrayBuffer);
+        let metadataBase64 = uint8ArrayToBase64(metadataArray);
+        transactionRecord.metadata = metadataBase64;
 
         let data = {
           id: transactionRecord.id,
-          accountId: transactionRecord.accountId,
-          initAccountState: transactionRecord.initAccountState,
-          finalAccountState: transactionRecord.finalAccountState,
-          inputNotes: transactionRecord.inputNotes,
-          outputNotes: transactionRecord.outputNotes,
+          metadata: transactionRecord.metadata,
           scriptRoot: transactionRecord.scriptRoot
             ? transactionRecord.scriptRoot
             : null,
@@ -143,18 +133,13 @@ export async function insertTransactionScript(scriptRoot, txScript) {
 
 export async function insertProvenTransactionData(
   transactionId,
-  accountId,
-  initAccountState,
-  finalAccountState,
-  inputNotes,
-  outputNotes,
+  metadata,
   scriptRoot,
   blockNum,
   committed
 ) {
   try {
-    let inputNotesBlob = new Blob([new Uint8Array(inputNotes)]);
-    let outputNotesBlob = new Blob([new Uint8Array(outputNotes)]);
+    let metadataBlob = new Blob([new Uint8Array(metadata)]);
     let scriptRootBase64 = null;
     if (scriptRoot !== null) {
       let scriptRootArray = new Uint8Array(scriptRoot);
@@ -163,11 +148,7 @@ export async function insertProvenTransactionData(
 
     const data = {
       id: transactionId,
-      accountId: accountId,
-      initAccountState: initAccountState,
-      finalAccountState: finalAccountState,
-      inputNotes: inputNotesBlob,
-      outputNotes: outputNotesBlob,
+      metadata: metadataBlob,
       scriptRoot: scriptRootBase64,
       blockNum: blockNum,
       commitHeight: committed ? committed : null,