+++ crates/rust-client/src/transaction/mod.rs
@@ -244,46 +244,80 @@ impl Deserializable for TransactionResult {
 // ================================================================================================
 
 /// Describes a transaction that has been executed and is being tracked on the Client.
-///
-/// Currently, the `commit_height` (and `committed` status) is set based on the height
-/// at which the transaction's output notes are committed.
 #[derive(Debug, Clone)]
 pub struct TransactionRecord {
+    /// Unique identifier for the transaction.
     pub id: TransactionId,
-    pub account_id: AccountId,
-    pub init_account_state: Digest,
-    pub final_account_state: Digest,
-    pub input_note_nullifiers: Vec<Digest>,
-    pub output_notes: OutputNotes,
-    pub transaction_script: Option<TransactionScript>,
-    pub block_num: BlockNumber,
-    pub transaction_status: TransactionStatus,
+    /// Metadata associated with the transaction.
+    pub metadata: TransactionMetadata,
+    /// Script associated with the transaction, if no script is provided, only note scripts are
+    /// executed.
+    pub script: Option<TransactionScript>,
+    /// Current status of the transaction.
+    pub status: TransactionStatus,
 }
 
 impl TransactionRecord {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         id: TransactionId,
-        account_id: AccountId,
-        init_account_state: Digest,
-        final_account_state: Digest,
-        input_note_nullifiers: Vec<Digest>,
-        output_notes: OutputNotes,
-        transaction_script: Option<TransactionScript>,
-        block_num: BlockNumber,
-        transaction_status: TransactionStatus,
+        metadata: TransactionMetadata,
+        script: Option<TransactionScript>,
+        status: TransactionStatus,
     ) -> TransactionRecord {
-        TransactionRecord {
-            id,
+        TransactionRecord { id, metadata, script, status }
+    }
+}
+
+/// Describes the metadata associated with a transaction.
+#[derive(Debug, Clone)]
+pub struct TransactionMetadata {
+    /// ID of the account that executed the transaction.
+    pub account_id: AccountId,
+    /// Initial state of the account before the transaction was executed.
+    pub init_account_state: Digest,
+    /// Final state of the account after the transaction was executed.
+    pub final_account_state: Digest,
+    /// Nullifiers of the input notes consumed in the transaction.
+    pub input_note_nullifiers: Vec<Digest>,
+    /// Output notes generated as a result of the transaction.
+    pub output_notes: OutputNotes,
+    /// Block number for the block against which the transaction was executed.
+    pub block_num: BlockNumber,
+    pub expiration_block_num: BlockNumber,
+}
+
+impl Serializable for TransactionMetadata {
+    fn write_into<W: ByteWriter>(&self, target: &mut W) {
+        self.account_id.write_into(target);
+        self.init_account_state.write_into(target);
+        self.final_account_state.write_into(target);
+        self.input_note_nullifiers.write_into(target);
+        self.output_notes.write_into(target);
+        self.block_num.write_into(target);
+        self.expiration_block_num.write_into(target);
+    }
+}
+
+impl Deserializable for TransactionMetadata {
+    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
+        let account_id = AccountId::read_from(source)?;
+        let init_account_state = Digest::read_from(source)?;
+        let final_account_state = Digest::read_from(source)?;
+        let input_note_nullifiers = Vec::<Digest>::read_from(source)?;
+        let output_notes = OutputNotes::read_from(source)?;
+        let block_num = BlockNumber::read_from(source)?;
+        let expiration_block_num = BlockNumber::read_from(source)?;
+
+        Ok(Self {
             account_id,
             init_account_state,
             final_account_state,
             input_note_nullifiers,
             output_notes,
-            transaction_script,
             block_num,
-            transaction_status,
-        }
+            expiration_block_num,
+        })
     }
 }